web-app/pyproject.toml
[project]
name = "vuls-web"
version = "0.1.0"
description = "Add your description here"
authors = [
  { name = "Paul Bargewell", email = "paul.bargewell@firebolt-group.com" },
]
requires-python = ">=3.11"
dependencies = [
  "aiofiles>=24.1.0",
  "alembic>=1.16.2",
  "celery>=5.5.3",
  "croniter>=3.0.3",
  "docker>=7.1.0",
  "fastapi>=0.115.14",
  "httpx>=0.27.0",
  "jinja2>=3.1.6",
  "argon2-cffi>=23.1.0",
  "bcrypt>=4.0.0",
  "psycopg[binary]>=3.1.0",
  "pydantic>=2.11.7",
  "pydantic-settings>=2.10.1",
  "python-crontab>=3.2.0",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.20",
  "redis[hiredis]>=5.0.0",
  "sqlalchemy>=2.0.41",
  "uvicorn[standard]>=0.34.3",
  "websockets>=12.0",
  "toml>=0.10.2",
  "tomli>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
